convert repository to factory pattern
